deep_research_app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
duckduckgo_search # Use latest stable
Flask==3.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.2
# Markdown package not strictly needed if markdown-it-py is used exclusively
markdown-it-py==3.0.0 # Primary markdown converter
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2 # For footnote plugin
mdurl==0.1.2
# openai # REMOVED - Was not used
pathspec==0.12.1
plumbum==1.9.0
ply==3.11
primp==0.14.0
proto-plus==1.26.1
protobuf>=4.25.3,<6.0.0dev # Version constraint for compatibility
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
tinycss2==1.4.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
# Gunicorn is recommended for production, but not strictly required for the app to run
gunicorn